requirements.txt
asn1crypto==0.23.0
-e git+git@github.com:zhexiao/hddep.git@e1345e71bd513598f2007cceb65b29e95ed94166#egg=audep
bcrypt==3.1.4
cffi==1.11.2
cryptography==2.1.1
Fabric3==1.13.1.post1
hddep==1.2.0
idna==2.6
paramiko==2.3.1
pyasn1==0.3.7
pycparser==2.18
PyNaCl==1.1.2
six==1.11.0
